chore: remove timeout in notify
