+++ src/main.rs
@@ -1,4 +1,5 @@
 use clap::Parser;
+use log::warn;
 use rustywatch::{
     args::{self, Args},
     logger, run,
@@ -14,7 +15,11 @@ async fn main() {
     let args = Args::parse();
 
     match Path::new(&args.config).exists() {
-        true => run::config(args).await.unwrap(),
-        false => run::cli(args).await.unwrap(),
+        true => run::config(args)
+            .await
+            .unwrap_or_else(|e: Box<dyn std::error::Error>| warn!("Error to execute: {}", e)),
+        false => run::cli(args)
+            .await
+            .unwrap_or_else(|e: Box<dyn std::error::Error>| warn!("Error to execute: {}", e)),
     }
 }
